refactor: move the components into same folder
